Refactor tests and mocks
modified:   src/fs-detail.js
modified:   test/fs-detail.test.js
modified:   test/fs-question.test.js
modified:   test/fs-result.test.js
renamed:    test/mocks/detailMock.js -> test/mocks/detail.js
new file:   test/mocks/question.js
renamed:    test/data.js -> test/mocks/result.js
modified:   test/util.js
