+++ test/mocks/detail.js
